@@ -1,19 +1,13 @@
-// swift-tools-version:5.10
+// swift-tools-version:6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let swiftSettings: Array<SwiftSetting> = [
-    .enableUpcomingFeature("ConciseMagicFile"),
+    .swiftLanguageMode(.v6),
     .enableUpcomingFeature("ExistentialAny"),
-    .enableUpcomingFeature("BareSlashRegexLiterals"),
-    .enableUpcomingFeature("DisableOutwardActorInference"),
-    .enableUpcomingFeature("IsolatedDefaultValues"),
-    .enableUpcomingFeature("DeprecateApplicationMain"),
-    .enableExperimentalFeature("StrictConcurrency"),
+    .enableUpcomingFeature("InternalImportsByDefault"),
     .enableExperimentalFeature("GlobalConcurrency"),
-//    .enableExperimentalFeature("AccessLevelOnImport"),
-//    .enableExperimentalFeature("VariadicGenerics"),
 ]
 
 let package = Package(
